Onlineshop/requirements.txt
Django==3.2.9
psycopg2-binary>=2.8
python-dotenv==0.19.2
djangorestframework==3.12.4

